- Changed some of the primary keys. - Changed some of the functionality of repository base. - Changed sql server with postgresql. - Initialized web layer with backend and frontend projects.

+++ DND.Middleware/Dto/Identity/ResetPasswordDto.cs
@@ -13,7 +13,7 @@ namespace DND.Middleware.Dto.Identity
         [StringLength(MaxLengths.PasswordText, MinimumLength = MinLengths.PasswordText)]
         public string ConfirmPassword { get; set; }
 
-        public long UserId { get; set; }
+        public int UserId { get; set; }
 
         [Required]
         public string Token { get; set; }
